pyproject.toml
[project]
name = "analize"
version = "0.1.0"
description = "Medical test analysis using LLM vision"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1",
    "google-genai>=1.51",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=6.0",
    "ruff>=0.14",
    "mypy>=1.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/analize"]

# Ruff - linting and formatting
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["analize"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Mypy - type checking
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
